remove casters

+++ src/components/amd_smi/amds_accessors.c
@@ -1249,22 +1249,22 @@ int access_amdsmi_gpu_metrics(int mode, void *arg) {
     event->value = metrics.pcie_nak_rcvd_count_acc;
     break;
   case 11:
-    event->value = (int64_t)metrics.accumulation_counter;
+    event->value = metrics.accumulation_counter;
     break;
   case 12:
-    event->value = (int64_t)metrics.prochot_residency_acc;
+    event->value = metrics.prochot_residency_acc;
     break;
   case 13:
-    event->value = (int64_t)metrics.ppt_residency_acc;
+    event->value = metrics.ppt_residency_acc;
     break;
   case 14:
-    event->value = (int64_t)metrics.socket_thm_residency_acc;
+    event->value = metrics.socket_thm_residency_acc;
     break;
   case 15:
-    event->value = (int64_t)metrics.vr_thm_residency_acc;
+    event->value = metrics.vr_thm_residency_acc;
     break;
   case 16:
-    event->value = (int64_t)metrics.hbm_thm_residency_acc;
+    event->value = metrics.hbm_thm_residency_acc;
     break;
   case 17:
     event->value = metrics.gfx_activity_acc;
